Move files pt. 3
