+++ packages/seroval/src/core/context/parser/async.ts
@@ -1,108 +1,115 @@
 /* eslint-disable prefer-destructuring */
-import type { BigIntTypedArrayValue, TypedArrayValue } from '../../types';
-import UnsupportedTypeError from '../UnsupportedTypeError';
-import assert from '../utils/assert';
+/* eslint-disable no-await-in-loop */
+import type { BigIntTypedArrayValue, TypedArrayValue } from '../../../types';
+import UnsupportedTypeError from '../../UnsupportedTypeError';
+import assert from '../../utils/assert';
 import {
-  createAggregateErrorNode,
-  createArrayBufferNode,
-  createArrayNode,
-  createBigIntNode,
-  createBigIntTypedArrayNode,
-  createBoxedNode,
-  createDataViewNode,
-  createDateNode,
-  createErrorNode,
-  createNumberNode,
   createPluginNode,
+  createDateNode,
   createRegExpNode,
-  createSetNode,
+  createArrayBufferNode,
+  createBigIntNode,
   createStringNode,
+  createNumberNode,
+  createArrayNode,
+  createBoxedNode,
   createTypedArrayNode,
-  createWKSymbolNode,
-} from '../base-primitives';
-import { BIGINT_FLAG, Feature } from '../compat';
-import type { WellKnownSymbols } from '../constants';
+  createBigIntTypedArrayNode,
+  createDataViewNode,
+  createErrorNode,
+  createSetNode,
+  createAggregateErrorNode,
+} from '../../base-primitives';
+import { BIGINT_FLAG, Feature } from '../../compat';
 import {
   SerovalNodeType,
   UNIVERSAL_SENTINEL,
-} from '../constants';
-import { createCustomEventOptions, createEventOptions } from '../utils/constructors';
+} from '../../constants';
 import {
-  FALSE_NODE,
+  createRequestOptions,
+  createResponseOptions,
+  createEventOptions,
+  createCustomEventOptions,
+} from '../../utils/constructors';
+import {
   NULL_NODE,
   TRUE_NODE,
+  FALSE_NODE,
   UNDEFINED_NODE,
-} from '../literals';
-import { iteratorToSequence } from '../utils/iterator-to-sequence';
-import type { BaseParserContextOptions } from '../parser-context';
-import { BaseParserContext } from '../parser-context';
-import { hasReferenceID } from '../reference';
-import { getErrorOptions } from '../utils/error';
-import { serializeString } from '../string';
-import { SerovalObjectRecordSpecialKey } from '../types';
+} from '../../literals';
+import { asyncIteratorToSequence, iteratorToSequence } from '../../utils/iterator-to-sequence';
+import { BaseParserContext } from '../parser';
+import promiseToResult from '../../utils/promise-to-result';
+import { getErrorOptions } from '../../utils/error';
+import { serializeString } from '../../string';
+import { SerovalObjectRecordSpecialKey } from '../../types';
 import type {
-  SerovalBoxedNode,
+  SerovalErrorNode,
   SerovalArrayNode,
+  SerovalBigIntTypedArrayNode,
+  SerovalBoxedNode,
   SerovalNode,
   SerovalNullConstructorNode,
   SerovalObjectNode,
   SerovalObjectRecordKey,
   SerovalObjectRecordNode,
-  SerovalErrorNode,
-  SerovalMapNode,
-  SerovalSetNode,
-  SerovalPluginNode,
+  SerovalPromiseNode,
+  SerovalTypedArrayNode,
   SerovalAggregateErrorNode,
+  SerovalBlobNode,
   SerovalCustomEventNode,
   SerovalEventNode,
-  SerovalHeadersNode,
-  SerovalPlainRecordNode,
+  SerovalFileNode,
   SerovalFormDataNode,
-  SerovalTypedArrayNode,
-  SerovalBigIntTypedArrayNode,
+  SerovalHeadersNode,
+  SerovalMapNode,
+  SerovalPlainRecordNode,
+  SerovalPluginNode,
+  SerovalRequestNode,
+  SerovalResponseNode,
+  SerovalSetNode,
   SerovalDataViewNode,
-} from '../types';
+} from '../../types';
 import {
-  createCustomEVentNode,
-  createDOMExceptionNode,
-  createEventNode,
   createURLNode,
   createURLSearchParamsNode,
-} from '../web-api';
+  createDOMExceptionNode,
+  createEventNode,
+  createCustomEVentNode,
+} from '../../web-api';
 
 type ObjectLikeNode =
   | SerovalObjectNode
-  | SerovalNullConstructorNode;
+  | SerovalNullConstructorNode
+  | SerovalPromiseNode;
 
-export type BaseSyncParserContextOptions = BaseParserContextOptions
-
-export default abstract class BaseSyncParserContext extends BaseParserContext {
-  protected parseItems(
+export default abstract class BaseAsyncParserContext extends BaseParserContext {
+  private async parseItems(
     current: unknown[],
-  ): SerovalNode[] {
+  ): Promise<SerovalNode[]> {
     const nodes = [];
     for (let i = 0, len = current.length; i < len; i++) {
       if (i in current) {
-        nodes[i] = this.parse(current[i]);
+        nodes[i] = await this.parse(current[i]);
       }
     }
     return nodes;
   }
 
-  protected parseArray(
+  private async parseArray(
     id: number,
     current: unknown[],
-  ): SerovalArrayNode {
+  ): Promise<SerovalArrayNode> {
     return createArrayNode(
       id,
       current,
-      this.parseItems(current),
+      await this.parseItems(current),
     );
   }
 
-  protected parseProperties(
+  private async parseProperties(
     properties: Record<string, unknown>,
-  ): SerovalObjectRecordNode {
+  ): Promise<SerovalObjectRecordNode> {
     const entries = Object.entries(properties);
     const keyNodes: SerovalObjectRecordKey[] = [];
     const valueNodes: SerovalNode[] = [];
@@ -112,13 +119,21 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
       i++
     ) {
       keyNodes.push(serializeString(entries[i][0]));
-      valueNodes.push(this.parse(entries[i][1]));
+      valueNodes.push(await this.parse(entries[i][1]));
     }
-    // Check special properties, symbols in this case
+    // Check special properties
     if (this.features & Feature.Symbol) {
       if (Symbol.iterator in properties) {
         keyNodes.push(SerovalObjectRecordSpecialKey.SymbolIterator);
-        valueNodes.push(this.parse(iteratorToSequence(properties as Iterable<unknown>)));
+        valueNodes.push(
+          await this.parse(iteratorToSequence(properties as Iterable<unknown>)),
+        );
+      }
+      if (Symbol.asyncIterator in properties) {
+        keyNodes.push(SerovalObjectRecordSpecialKey.SymbolAsyncIterator);
+        valueNodes.push(
+          await this.parse(await asyncIteratorToSequence(properties as AsyncIterable<unknown>)),
+        );
       }
     }
     return {
@@ -128,84 +143,84 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     };
   }
 
-  protected parsePlainObject(
+  private async parsePlainObject(
     id: number,
     current: Record<string, unknown>,
     empty: boolean,
-  ): ObjectLikeNode {
+  ): Promise<ObjectLikeNode> {
     return this.createObjectNode(
       id,
       current,
       empty,
-      this.parseProperties(current),
+      await this.parseProperties(current),
     );
   }
 
-  protected parseBoxed(
+  private async parseBoxed(
     id: number,
     current: object,
-  ): SerovalBoxedNode {
-    return createBoxedNode(id, this.parse(current.valueOf()));
+  ): Promise<SerovalBoxedNode> {
+    return createBoxedNode(id, await this.parse(current.valueOf()));
   }
 
-  protected parseTypedArray(
+  private async parseTypedArray(
     id: number,
     current: TypedArrayValue,
-  ): SerovalTypedArrayNode {
-    return createTypedArrayNode(id, current, this.parse(current.buffer));
+  ): Promise<SerovalTypedArrayNode> {
+    return createTypedArrayNode(id, current, await this.parse(current.buffer));
   }
 
-  protected parseBigIntTypedArray(
+  private async parseBigIntTypedArray(
     id: number,
     current: BigIntTypedArrayValue,
-  ): SerovalBigIntTypedArrayNode {
-    return createBigIntTypedArrayNode(id, current, this.parse(current.buffer));
+  ): Promise<SerovalBigIntTypedArrayNode> {
+    return createBigIntTypedArrayNode(id, current, await this.parse(current.buffer));
   }
 
-  protected parseDataView(
+  private async parseDataView(
     id: number,
     current: DataView,
-  ): SerovalDataViewNode {
-    return createDataViewNode(id, current, this.parse(current.buffer));
+  ): Promise<SerovalDataViewNode> {
+    return createDataViewNode(id, current, await this.parse(current.buffer));
   }
 
-  protected parseError(
+  private async parseError(
     id: number,
     current: Error,
-  ): SerovalErrorNode {
+  ): Promise<SerovalErrorNode> {
     const options = getErrorOptions(current, this.features);
     return createErrorNode(
       id,
       current,
       options
-        ? this.parseProperties(options)
+        ? await this.parseProperties(options)
         : undefined,
     );
   }
 
-  protected parseAggregateError(
+  private async parseAggregateError(
     id: number,
     current: AggregateError,
-  ): SerovalAggregateErrorNode {
+  ): Promise<SerovalAggregateErrorNode> {
     const options = getErrorOptions(current, this.features);
     return createAggregateErrorNode(
       id,
       current,
       options
-        ? this.parseProperties(options)
+        ? await this.parseProperties(options)
         : undefined,
     );
   }
 
-  protected parseMap(
+  private async parseMap(
     id: number,
     current: Map<unknown, unknown>,
-  ): SerovalMapNode {
+  ): Promise<SerovalMapNode> {
     const keyNodes: SerovalNode[] = [];
     const valueNodes: SerovalNode[] = [];
     for (const [key, value] of current.entries()) {
-      keyNodes.push(this.parse(key));
-      valueNodes.push(this.parse(value));
+      keyNodes.push(await this.parse(key));
+      valueNodes.push(await this.parse(value));
     }
     return this.createMapNode(
       id,
@@ -215,26 +230,68 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     );
   }
 
-  protected parseSet(
+  private async parseSet(
     id: number,
     current: Set<unknown>,
-  ): SerovalSetNode {
+  ): Promise<SerovalSetNode> {
     const items: SerovalNode[] = [];
     for (const item of current.keys()) {
-      items.push(this.parse(item));
+      items.push(await this.parse(item));
     }
     return createSetNode(id, current.size, items);
   }
 
-  protected parsePlainProperties(
+  private async parseBlob(
+    id: number,
+    current: Blob,
+  ): Promise<SerovalBlobNode> {
+    return {
+      t: SerovalNodeType.Blob,
+      i: id,
+      s: undefined,
+      l: undefined,
+      c: serializeString(current.type),
+      m: undefined,
+      p: undefined,
+      e: undefined,
+      f: await this.parse(await current.arrayBuffer()),
+      a: undefined,
+      b: undefined,
+      o: undefined,
+      x: undefined,
+    };
+  }
+
+  private async parseFile(
+    id: number,
+    current: File,
+  ): Promise<SerovalFileNode> {
+    return {
+      t: SerovalNodeType.File,
+      i: id,
+      s: undefined,
+      l: undefined,
+      c: serializeString(current.type),
+      m: serializeString(current.name),
+      p: undefined,
+      e: undefined,
+      f: await this.parse(await current.arrayBuffer()),
+      a: undefined,
+      b: current.lastModified,
+      o: undefined,
+      x: undefined,
+    };
+  }
+
+  protected async parsePlainProperties(
     entries: [key: string, value: unknown][],
-  ): SerovalPlainRecordNode {
+  ): Promise<SerovalPlainRecordNode> {
     const size = entries.length;
     const keyNodes: string[] = [];
     const valueNodes: SerovalNode[] = [];
     for (let i = 0; i < size; i++) {
       keyNodes.push(serializeString(entries[i][0]));
-      valueNodes.push(this.parse(entries[i][1]));
+      valueNodes.push(await this.parse(entries[i][1]));
     }
     return {
       k: keyNodes,
@@ -243,10 +300,10 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     };
   }
 
-  protected parseHeaders(
+  private async parseHeaders(
     id: number,
     current: Headers,
-  ): SerovalHeadersNode {
+  ): Promise<SerovalHeadersNode> {
     const items: [key: string, value: unknown][] = [];
     current.forEach((value, key) => {
       items.push([key, value]);
@@ -259,7 +316,7 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
       c: undefined,
       m: undefined,
       p: undefined,
-      e: this.parsePlainProperties(items),
+      e: await this.parsePlainProperties(items),
       a: undefined,
       f: undefined,
       b: undefined,
@@ -268,11 +325,11 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     };
   }
 
-  protected parseFormData(
+  private async parseFormData(
     id: number,
     current: FormData,
-  ): SerovalFormDataNode {
-    const items: [key: string, value: unknown][] = [];
+  ): Promise<SerovalFormDataNode> {
+    const items: [key: string, value: FormDataEntryValue][] = [];
     current.forEach((value, key) => {
       items.push([key, value]);
     });
@@ -284,7 +341,7 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
       c: undefined,
       m: undefined,
       p: undefined,
-      e: this.parsePlainProperties(items),
+      e: await this.parsePlainProperties(items),
       a: undefined,
       f: undefined,
       b: undefined,
@@ -293,33 +350,108 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     };
   }
 
-  protected parseEvent(
+  private async parseRequest(
+    id: number,
+    current: Request,
+  ): Promise<SerovalRequestNode> {
+    return {
+      t: SerovalNodeType.Request,
+      i: id,
+      s: serializeString(current.url),
+      l: undefined,
+      c: undefined,
+      m: undefined,
+      p: undefined,
+      e: undefined,
+      f: await this.parse(
+        createRequestOptions(current, current.body ? await current.clone().arrayBuffer() : null),
+      ),
+      a: undefined,
+      b: undefined,
+      o: undefined,
+      x: undefined,
+    };
+  }
+
+  private async parseResponse(
+    id: number,
+    current: Response,
+  ): Promise<SerovalResponseNode> {
+    return {
+      t: SerovalNodeType.Response,
+      i: id,
+      s: undefined,
+      l: undefined,
+      c: undefined,
+      m: undefined,
+      p: undefined,
+      e: undefined,
+      f: undefined,
+      a: [
+        current.body
+          ? await this.parse(await current.clone().arrayBuffer())
+          : NULL_NODE,
+        await this.parse(createResponseOptions(current)),
+      ],
+      b: undefined,
+      o: undefined,
+      x: undefined,
+    };
+  }
+
+  private async parseEvent(
     id: number,
     current: Event,
-  ): SerovalEventNode {
-    return createEventNode(id, current.type, this.parse(createEventOptions(current)));
+  ): Promise<SerovalEventNode> {
+    return createEventNode(id, current.type, await this.parse(createEventOptions(current)));
   }
 
-  protected parseCustomEvent(
+  private async parseCustomEvent(
     id: number,
     current: CustomEvent,
-  ): SerovalCustomEventNode {
-    return createCustomEVentNode(id, current.type, this.parse(createCustomEventOptions(current)));
+  ): Promise<SerovalCustomEventNode> {
+    return createCustomEVentNode(
+      id,
+      current.type,
+      await this.parse(createCustomEventOptions(current)),
+    );
   }
 
-  protected parsePlugin(
+  private async parsePromise(
+    id: number,
+    current: Promise<unknown>,
+  ): Promise<SerovalPromiseNode> {
+    const [status, result] = await promiseToResult(current);
+    return {
+      t: SerovalNodeType.Promise,
+      i: id,
+      s: status,
+      l: undefined,
+      c: undefined,
+      m: undefined,
+      p: undefined,
+      e: undefined,
+      a: undefined,
+      f: await this.parse(result),
+      b: undefined,
+      o: undefined,
+      x: undefined,
+    };
+  }
+
+  private async parsePlugin(
     id: number,
     current: unknown,
-  ): SerovalPluginNode | undefined {
+  ): Promise<SerovalPluginNode | undefined> {
     const currentPlugins = this.plugins;
     if (currentPlugins) {
       for (let i = 0, len = currentPlugins.length; i < len; i++) {
         const plugin = currentPlugins[i];
-        if (plugin.parse.sync && plugin.test(current)) {
+        if (plugin.parse.async && plugin.test(current)) {
           return createPluginNode(
             id,
             plugin.tag,
-            plugin.parse.sync(current, this, {
+            await plugin.parse.async(current, this, {
               id,
             }),
           );
@@ -329,10 +461,10 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     return undefined;
   }
 
-  protected parseObject(
+  private async parseObject(
     id: number,
     current: object,
-  ): SerovalNode {
+  ): Promise<SerovalNode> {
     if (Array.isArray(current)) {
       return this.parseArray(id, current);
     }
@@ -371,6 +503,13 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
         break;
     }
     const currentFeatures = this.features;
+    // Promises
+    if (
+      (currentFeatures & Feature.Promise)
+      && (currentClass === Promise || current instanceof Promise)
+    ) {
+      return this.parsePromise(id, current as unknown as Promise<unknown>);
+    }
     // Typed Arrays
     if (currentFeatures & Feature.TypedArray) {
       switch (currentClass) {
@@ -422,10 +561,18 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
           return createURLNode(id, current as unknown as URL);
         case (typeof URLSearchParams !== 'undefined' ? URLSearchParams : UNIVERSAL_SENTINEL):
           return createURLSearchParamsNode(id, current as unknown as URLSearchParams);
+        case (typeof Blob !== 'undefined' ? Blob : UNIVERSAL_SENTINEL):
+          return this.parseBlob(id, current as unknown as Blob);
+        case (typeof File !== 'undefined' ? File : UNIVERSAL_SENTINEL):
+          return this.parseFile(id, current as unknown as File);
         case (typeof Headers !== 'undefined' ? Headers : UNIVERSAL_SENTINEL):
           return this.parseHeaders(id, current as unknown as Headers);
         case (typeof FormData !== 'undefined' ? FormData : UNIVERSAL_SENTINEL):
           return this.parseFormData(id, current as unknown as FormData);
+        case (typeof Request !== 'undefined' ? Request : UNIVERSAL_SENTINEL):
+          return this.parseRequest(id, current as unknown as Request);
+        case (typeof Response !== 'undefined' ? Response : UNIVERSAL_SENTINEL):
+          return this.parseResponse(id, current as unknown as Response);
         case (typeof Event !== 'undefined' ? Event : UNIVERSAL_SENTINEL):
           return this.parseEvent(id, current as unknown as Event);
         case (typeof CustomEvent !== 'undefined' ? CustomEvent : UNIVERSAL_SENTINEL):
@@ -436,7 +583,7 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
           break;
       }
     }
-    const parsed = this.parsePlugin(id, current);
+    const parsed = await this.parsePlugin(id, current);
     if (parsed) {
       return parsed;
     }
@@ -456,14 +603,14 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
     // despite existing
     if (
       currentFeatures & Feature.Symbol
-      && Symbol.iterator in current
+      && (Symbol.iterator in current || Symbol.asyncIterator in current)
     ) {
       return this.parsePlainObject(id, current, !!currentClass);
     }
     throw new UnsupportedTypeError(current);
   }
 
-  parse<T>(current: T): SerovalNode {
+  async parse<T>(current: T): Promise<SerovalNode> {
     switch (current) {
       case true: return TRUE_NODE;
       case false: return FALSE_NODE;
@@ -481,14 +628,10 @@ export default abstract class BaseSyncParserContext extends BaseParserContext {
         const id = this.getReference(current);
         return typeof id === 'number' ? this.parseObject(id, current as object) : id;
       }
-      case 'symbol': {
-        assert(this.features & Feature.Symbol, new UnsupportedTypeError(current));
-        const id = this.getReference(current);
-        return typeof id === 'number' ? createWKSymbolNode(id, current as WellKnownSymbols) : id;
-      }
+      case 'symbol':
+        return this.parseWKSymbol(current);
       case 'function':
-        assert(hasReferenceID(current), new Error('Cannot serialize function without reference ID.'));
-        return this.getStrictReference(current);
+        return this.parseFunction(current);
       default:
         throw new UnsupportedTypeError(current);
     }
