+++ packages/seroval/src/core/context/parser/sync.ts
@@ -1,115 +1,108 @@
 /* eslint-disable prefer-destructuring */
-/* eslint-disable no-await-in-loop */
-import type { BigIntTypedArrayValue, TypedArrayValue } from '../../types';
-import UnsupportedTypeError from '../UnsupportedTypeError';
-import assert from '../utils/assert';
+import type { BigIntTypedArrayValue, TypedArrayValue } from '../../../types';
+import UnsupportedTypeError from '../../UnsupportedTypeError';
+import assert from '../../utils/assert';
 import {
-  createPluginNode,
-  createDateNode,
-  createRegExpNode,
-  createArrayBufferNode,
-  createBigIntNode,
-  createStringNode,
-  createNumberNode,
-  createArrayNode,
-  createBoxedNode,
-  createTypedArrayNode,
-  createBigIntTypedArrayNode,
-  createDataViewNode,
-  createErrorNode,
-  createSetNode,
   createAggregateErrorNode,
-} from '../base-primitives';
-import { BIGINT_FLAG, Feature } from '../compat';
+  createArrayBufferNode,
+  createArrayNode,
+  createBigIntNode,
+  createBigIntTypedArrayNode,
+  createBoxedNode,
+  createDataViewNode,
+  createDateNode,
+  createErrorNode,
+  createNumberNode,
+  createPluginNode,
+  createRegExpNode,
+  createSetNode,
+  createStringNode,
+  createTypedArrayNode,
+  createWKSymbolNode,
+} from '../../base-primitives';
+import { BIGINT_FLAG, Feature } from '../../compat';
+import type { WellKnownSymbols } from '../../constants';
 import {
   SerovalNodeType,
   UNIVERSAL_SENTINEL,
-} from '../constants';
+} from '../../constants';
+import { createCustomEventOptions, createEventOptions } from '../../utils/constructors';
 import {
-  createRequestOptions,
-  createResponseOptions,
-  createEventOptions,
-  createCustomEventOptions,
-} from '../utils/constructors';
-import {
+  FALSE_NODE,
   NULL_NODE,
   TRUE_NODE,
-  FALSE_NODE,
   UNDEFINED_NODE,
-} from '../literals';
-import { asyncIteratorToSequence, iteratorToSequence } from '../utils/iterator-to-sequence';
-import { BaseParserContext } from '../parser-context';
-import promiseToResult from '../utils/promise-to-result';
-import { getErrorOptions } from '../utils/error';
-import { serializeString } from '../string';
-import { SerovalObjectRecordSpecialKey } from '../types';
+} from '../../literals';
+import { iteratorToSequence } from '../../utils/iterator-to-sequence';
+import type { BaseParserContextOptions } from '../parser';
+import { BaseParserContext } from '../parser';
+import { hasReferenceID } from '../../reference';
+import { getErrorOptions } from '../../utils/error';
+import { serializeString } from '../../string';
+import { SerovalObjectRecordSpecialKey } from '../../types';
 import type {
-  SerovalErrorNode,
-  SerovalArrayNode,
-  SerovalBigIntTypedArrayNode,
   SerovalBoxedNode,
+  SerovalArrayNode,
   SerovalNode,
   SerovalNullConstructorNode,
   SerovalObjectNode,
   SerovalObjectRecordKey,
   SerovalObjectRecordNode,
-  SerovalPromiseNode,
-  SerovalTypedArrayNode,
+  SerovalErrorNode,
+  SerovalMapNode,
+  SerovalSetNode,
+  SerovalPluginNode,
   SerovalAggregateErrorNode,
-  SerovalBlobNode,
   SerovalCustomEventNode,
   SerovalEventNode,
-  SerovalFileNode,
-  SerovalFormDataNode,
   SerovalHeadersNode,
-  SerovalMapNode,
   SerovalPlainRecordNode,
-  SerovalPluginNode,
-  SerovalRequestNode,
-  SerovalResponseNode,
-  SerovalSetNode,
+  SerovalFormDataNode,
+  SerovalTypedArrayNode,
+  SerovalBigIntTypedArrayNode,
   SerovalDataViewNode,
-} from '../types';
+} from '../../types';
 import {
-  createURLNode,
-  createURLSearchParamsNode,
+  createCustomEVentNode,
   createDOMExceptionNode,
   createEventNode,
-  createCustomEVentNode,
-} from '../web-api';
+  createURLNode,
+  createURLSearchParamsNode,
+} from '../../web-api';
 
 type ObjectLikeNode =
   | SerovalObjectNode
-  | SerovalNullConstructorNode
-  | SerovalPromiseNode;
+  | SerovalNullConstructorNode;
 
-export default abstract class BaseAsyncParserContext extends BaseParserContext {
-  private async parseItems(
+export type BaseSyncParserContextOptions = BaseParserContextOptions
+
+export default abstract class BaseSyncParserContext extends BaseParserContext {
+  protected parseItems(
     current: unknown[],
-  ): Promise<SerovalNode[]> {
+  ): SerovalNode[] {
     const nodes = [];
     for (let i = 0, len = current.length; i < len; i++) {
       if (i in current) {
-        nodes[i] = await this.parse(current[i]);
+        nodes[i] = this.parse(current[i]);
       }
     }
     return nodes;
   }
 
-  private async parseArray(
+  protected parseArray(
     id: number,
     current: unknown[],
-  ): Promise<SerovalArrayNode> {
+  ): SerovalArrayNode {
     return createArrayNode(
       id,
       current,
-      await this.parseItems(current),
+      this.parseItems(current),
     );
   }
 
-  private async parseProperties(
+  protected parseProperties(
     properties: Record<string, unknown>,
-  ): Promise<SerovalObjectRecordNode> {
+  ): SerovalObjectRecordNode {
     const entries = Object.entries(properties);
     const keyNodes: SerovalObjectRecordKey[] = [];
     const valueNodes: SerovalNode[] = [];
@@ -119,21 +112,13 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
       i++
     ) {
       keyNodes.push(serializeString(entries[i][0]));
-      valueNodes.push(await this.parse(entries[i][1]));
+      valueNodes.push(this.parse(entries[i][1]));
     }
-    // Check special properties
+    // Check special properties, symbols in this case
     if (this.features & Feature.Symbol) {
       if (Symbol.iterator in properties) {
         keyNodes.push(SerovalObjectRecordSpecialKey.SymbolIterator);
-        valueNodes.push(
-          await this.parse(iteratorToSequence(properties as Iterable<unknown>)),
-        );
-      }
-      if (Symbol.asyncIterator in properties) {
-        keyNodes.push(SerovalObjectRecordSpecialKey.SymbolAsyncIterator);
-        valueNodes.push(
-          await this.parse(await asyncIteratorToSequence(properties as AsyncIterable<unknown>)),
-        );
+        valueNodes.push(this.parse(iteratorToSequence(properties as Iterable<unknown>)));
       }
     }
     return {
@@ -143,84 +128,84 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     };
   }
 
-  private async parsePlainObject(
+  protected parsePlainObject(
     id: number,
     current: Record<string, unknown>,
     empty: boolean,
-  ): Promise<ObjectLikeNode> {
+  ): ObjectLikeNode {
     return this.createObjectNode(
       id,
       current,
       empty,
-      await this.parseProperties(current),
+      this.parseProperties(current),
     );
   }
 
-  private async parseBoxed(
+  protected parseBoxed(
     id: number,
     current: object,
-  ): Promise<SerovalBoxedNode> {
-    return createBoxedNode(id, await this.parse(current.valueOf()));
+  ): SerovalBoxedNode {
+    return createBoxedNode(id, this.parse(current.valueOf()));
   }
 
-  private async parseTypedArray(
+  protected parseTypedArray(
     id: number,
     current: TypedArrayValue,
-  ): Promise<SerovalTypedArrayNode> {
-    return createTypedArrayNode(id, current, await this.parse(current.buffer));
+  ): SerovalTypedArrayNode {
+    return createTypedArrayNode(id, current, this.parse(current.buffer));
   }
 
-  private async parseBigIntTypedArray(
+  protected parseBigIntTypedArray(
     id: number,
     current: BigIntTypedArrayValue,
-  ): Promise<SerovalBigIntTypedArrayNode> {
-    return createBigIntTypedArrayNode(id, current, await this.parse(current.buffer));
+  ): SerovalBigIntTypedArrayNode {
+    return createBigIntTypedArrayNode(id, current, this.parse(current.buffer));
   }
 
-  private async parseDataView(
+  protected parseDataView(
     id: number,
     current: DataView,
-  ): Promise<SerovalDataViewNode> {
-    return createDataViewNode(id, current, await this.parse(current.buffer));
+  ): SerovalDataViewNode {
+    return createDataViewNode(id, current, this.parse(current.buffer));
   }
 
-  private async parseError(
+  protected parseError(
     id: number,
     current: Error,
-  ): Promise<SerovalErrorNode> {
+  ): SerovalErrorNode {
     const options = getErrorOptions(current, this.features);
     return createErrorNode(
       id,
       current,
       options
-        ? await this.parseProperties(options)
+        ? this.parseProperties(options)
         : undefined,
     );
   }
 
-  private async parseAggregateError(
+  protected parseAggregateError(
     id: number,
     current: AggregateError,
-  ): Promise<SerovalAggregateErrorNode> {
+  ): SerovalAggregateErrorNode {
     const options = getErrorOptions(current, this.features);
     return createAggregateErrorNode(
       id,
       current,
       options
-        ? await this.parseProperties(options)
+        ? this.parseProperties(options)
         : undefined,
     );
   }
 
-  private async parseMap(
+  protected parseMap(
     id: number,
     current: Map<unknown, unknown>,
-  ): Promise<SerovalMapNode> {
+  ): SerovalMapNode {
     const keyNodes: SerovalNode[] = [];
     const valueNodes: SerovalNode[] = [];
     for (const [key, value] of current.entries()) {
-      keyNodes.push(await this.parse(key));
-      valueNodes.push(await this.parse(value));
+      keyNodes.push(this.parse(key));
+      valueNodes.push(this.parse(value));
     }
     return this.createMapNode(
       id,
@@ -230,68 +215,26 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     );
   }
 
-  private async parseSet(
+  protected parseSet(
     id: number,
     current: Set<unknown>,
-  ): Promise<SerovalSetNode> {
+  ): SerovalSetNode {
     const items: SerovalNode[] = [];
     for (const item of current.keys()) {
-      items.push(await this.parse(item));
+      items.push(this.parse(item));
     }
     return createSetNode(id, current.size, items);
   }
 
-  private async parseBlob(
-    id: number,
-    current: Blob,
-  ): Promise<SerovalBlobNode> {
-    return {
-      t: SerovalNodeType.Blob,
-      i: id,
-      s: undefined,
-      l: undefined,
-      c: serializeString(current.type),
-      m: undefined,
-      p: undefined,
-      e: undefined,
-      f: await this.parse(await current.arrayBuffer()),
-      a: undefined,
-      b: undefined,
-      o: undefined,
-      x: undefined,
-    };
-  }
-
-  private async parseFile(
-    id: number,
-    current: File,
-  ): Promise<SerovalFileNode> {
-    return {
-      t: SerovalNodeType.File,
-      i: id,
-      s: undefined,
-      l: undefined,
-      c: serializeString(current.type),
-      m: serializeString(current.name),
-      p: undefined,
-      e: undefined,
-      f: await this.parse(await current.arrayBuffer()),
-      a: undefined,
-      b: current.lastModified,
-      o: undefined,
-      x: undefined,
-    };
-  }
-
-  protected async parsePlainProperties(
+  protected parsePlainProperties(
     entries: [key: string, value: unknown][],
-  ): Promise<SerovalPlainRecordNode> {
+  ): SerovalPlainRecordNode {
     const size = entries.length;
     const keyNodes: string[] = [];
     const valueNodes: SerovalNode[] = [];
     for (let i = 0; i < size; i++) {
       keyNodes.push(serializeString(entries[i][0]));
-      valueNodes.push(await this.parse(entries[i][1]));
+      valueNodes.push(this.parse(entries[i][1]));
     }
     return {
       k: keyNodes,
@@ -300,10 +243,10 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     };
   }
 
-  private async parseHeaders(
+  protected parseHeaders(
     id: number,
     current: Headers,
-  ): Promise<SerovalHeadersNode> {
+  ): SerovalHeadersNode {
     const items: [key: string, value: unknown][] = [];
     current.forEach((value, key) => {
       items.push([key, value]);
@@ -316,7 +259,7 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
       c: undefined,
       m: undefined,
       p: undefined,
-      e: await this.parsePlainProperties(items),
+      e: this.parsePlainProperties(items),
       a: undefined,
       f: undefined,
       b: undefined,
@@ -325,11 +268,11 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     };
   }
 
-  private async parseFormData(
+  protected parseFormData(
     id: number,
     current: FormData,
-  ): Promise<SerovalFormDataNode> {
-    const items: [key: string, value: FormDataEntryValue][] = [];
+  ): SerovalFormDataNode {
+    const items: [key: string, value: unknown][] = [];
     current.forEach((value, key) => {
       items.push([key, value]);
     });
@@ -341,7 +284,7 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
       c: undefined,
       m: undefined,
       p: undefined,
-      e: await this.parsePlainProperties(items),
+      e: this.parsePlainProperties(items),
       a: undefined,
       f: undefined,
       b: undefined,
@@ -350,108 +293,33 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     };
   }
 
-  private async parseRequest(
-    id: number,
-    current: Request,
-  ): Promise<SerovalRequestNode> {
-    return {
-      t: SerovalNodeType.Request,
-      i: id,
-      s: serializeString(current.url),
-      l: undefined,
-      c: undefined,
-      m: undefined,
-      p: undefined,
-      e: undefined,
-      f: await this.parse(
-        createRequestOptions(current, current.body ? await current.clone().arrayBuffer() : null),
-      ),
-      a: undefined,
-      b: undefined,
-      o: undefined,
-      x: undefined,
-    };
-  }
-
-  private async parseResponse(
-    id: number,
-    current: Response,
-  ): Promise<SerovalResponseNode> {
-    return {
-      t: SerovalNodeType.Response,
-      i: id,
-      s: undefined,
-      l: undefined,
-      c: undefined,
-      m: undefined,
-      p: undefined,
-      e: undefined,
-      f: undefined,
-      a: [
-        current.body
-          ? await this.parse(await current.clone().arrayBuffer())
-          : NULL_NODE,
-        await this.parse(createResponseOptions(current)),
-      ],
-      b: undefined,
-      o: undefined,
-      x: undefined,
-    };
-  }
-
-  private async parseEvent(
+  protected parseEvent(
     id: number,
     current: Event,
-  ): Promise<SerovalEventNode> {
-    return createEventNode(id, current.type, await this.parse(createEventOptions(current)));
+  ): SerovalEventNode {
+    return createEventNode(id, current.type, this.parse(createEventOptions(current)));
   }
 
-  private async parseCustomEvent(
+  protected parseCustomEvent(
     id: number,
     current: CustomEvent,
-  ): Promise<SerovalCustomEventNode> {
-    return createCustomEVentNode(
-      id,
-      current.type,
-      await this.parse(createCustomEventOptions(current)),
-    );
+  ): SerovalCustomEventNode {
+    return createCustomEVentNode(id, current.type, this.parse(createCustomEventOptions(current)));
   }
 
-  private async parsePromise(
-    id: number,
-    current: Promise<unknown>,
-  ): Promise<SerovalPromiseNode> {
-    const [status, result] = await promiseToResult(current);
-    return {
-      t: SerovalNodeType.Promise,
-      i: id,
-      s: status,
-      l: undefined,
-      c: undefined,
-      m: undefined,
-      p: undefined,
-      e: undefined,
-      a: undefined,
-      f: await this.parse(result),
-      b: undefined,
-      o: undefined,
-      x: undefined,
-    };
-  }
-
-  private async parsePlugin(
+  protected parsePlugin(
     id: number,
     current: unknown,
-  ): Promise<SerovalPluginNode | undefined> {
+  ): SerovalPluginNode | undefined {
     const currentPlugins = this.plugins;
     if (currentPlugins) {
       for (let i = 0, len = currentPlugins.length; i < len; i++) {
         const plugin = currentPlugins[i];
-        if (plugin.parse.async && plugin.test(current)) {
+        if (plugin.parse.sync && plugin.test(current)) {
           return createPluginNode(
             id,
             plugin.tag,
-            await plugin.parse.async(current, this, {
+            plugin.parse.sync(current, this, {
               id,
             }),
           );
@@ -461,10 +329,10 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     return undefined;
   }
 
-  private async parseObject(
+  protected parseObject(
     id: number,
     current: object,
-  ): Promise<SerovalNode> {
+  ): SerovalNode {
     if (Array.isArray(current)) {
       return this.parseArray(id, current);
     }
@@ -503,13 +371,6 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
         break;
     }
     const currentFeatures = this.features;
-    // Promises
-    if (
-      (currentFeatures & Feature.Promise)
-      && (currentClass === Promise || current instanceof Promise)
-    ) {
-      return this.parsePromise(id, current as unknown as Promise<unknown>);
-    }
     // Typed Arrays
     if (currentFeatures & Feature.TypedArray) {
       switch (currentClass) {
@@ -561,18 +422,10 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
           return createURLNode(id, current as unknown as URL);
         case (typeof URLSearchParams !== 'undefined' ? URLSearchParams : UNIVERSAL_SENTINEL):
           return createURLSearchParamsNode(id, current as unknown as URLSearchParams);
-        case (typeof Blob !== 'undefined' ? Blob : UNIVERSAL_SENTINEL):
-          return this.parseBlob(id, current as unknown as Blob);
-        case (typeof File !== 'undefined' ? File : UNIVERSAL_SENTINEL):
-          return this.parseFile(id, current as unknown as File);
         case (typeof Headers !== 'undefined' ? Headers : UNIVERSAL_SENTINEL):
           return this.parseHeaders(id, current as unknown as Headers);
         case (typeof FormData !== 'undefined' ? FormData : UNIVERSAL_SENTINEL):
           return this.parseFormData(id, current as unknown as FormData);
-        case (typeof Request !== 'undefined' ? Request : UNIVERSAL_SENTINEL):
-          return this.parseRequest(id, current as unknown as Request);
-        case (typeof Response !== 'undefined' ? Response : UNIVERSAL_SENTINEL):
-          return this.parseResponse(id, current as unknown as Response);
         case (typeof Event !== 'undefined' ? Event : UNIVERSAL_SENTINEL):
           return this.parseEvent(id, current as unknown as Event);
         case (typeof CustomEvent !== 'undefined' ? CustomEvent : UNIVERSAL_SENTINEL):
@@ -583,7 +436,7 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
           break;
       }
     }
-    const parsed = await this.parsePlugin(id, current);
+    const parsed = this.parsePlugin(id, current);
     if (parsed) {
       return parsed;
     }
@@ -603,14 +456,14 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
     // despite existing
     if (
       currentFeatures & Feature.Symbol
-      && (Symbol.iterator in current || Symbol.asyncIterator in current)
+      && Symbol.iterator in current
     ) {
       return this.parsePlainObject(id, current, !!currentClass);
     }
     throw new UnsupportedTypeError(current);
   }
 
-  async parse<T>(current: T): Promise<SerovalNode> {
+  parse<T>(current: T): SerovalNode {
     switch (current) {
       case true: return TRUE_NODE;
       case false: return FALSE_NODE;
@@ -628,10 +481,14 @@ export default abstract class BaseAsyncParserContext extends BaseParserContext {
         const id = this.getReference(current);
         return typeof id === 'number' ? this.parseObject(id, current as object) : id;
       }
-      case 'symbol':
-        return this.parseWKSymbol(current);
+      case 'symbol': {
+        assert(this.features & Feature.Symbol, new UnsupportedTypeError(current));
+        const id = this.getReference(current);
+        return typeof id === 'number' ? createWKSymbolNode(id, current as WellKnownSymbols) : id;
+      }
       case 'function':
-        return this.parseFunction(current);
+        assert(hasReferenceID(current), new Error('Cannot serialize function without reference ID.'));
+        return this.getStrictReference(current);
       default:
         throw new UnsupportedTypeError(current);
     }
