+++ packages/seroval/src/core/cross/cross-parser.ts
@@ -1,4 +1,4 @@
-import type { BaseParserContextOptions } from '../parser-context';
+import type { BaseParserContextOptions } from '../context/parser';
 
 export type CrossParserContextOptions = BaseParserContextOptions
