+++ packages/seroval/src/core/cross/stream.ts
@@ -1,5 +1,5 @@
-import type { BaseStreamParserContextOptions } from '../base-parser/stream';
-import BaseStreamParserContext from '../base-parser/stream';
+import type { BaseStreamParserContextOptions } from '../context/parser/stream';
+import BaseStreamParserContext from '../context/parser/stream';
 import type { SerovalMode } from '../plugin';
 
 export type CrossStreamParserContextOptions = BaseStreamParserContextOptions
